requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
blis==1.2.0
Brotli==1.1.0
catalogue==2.0.10
celery==5.4.0
certifi
cffi==1.17.1
chardet
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
cssselect2==0.7.0
cymem==2.0.11
Django==5.1.2
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-crontab==0.7.1
django-extensions==3.2.3
django-simple-history==3.7.0
django-tenants==3.7.0
django-timezone-field==7.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
filelock==3.17.0
fonttools==4.55.7
fsspec==2025.3.0
huggingface-hub==0.29.2
idna
Jinja2==3.1.6
joblib==1.4.2
kombu==5.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
num2words==0.5.13
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
preshed==3.0.9
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.2
pydyf==0.11.0
Pygments==2.19.1
pyparsing==3.2.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scipy==1.15.2
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sqlparse==0.5.1
srsly==2.5.1

sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==63.1
webencodings==0.5.1
wheel==0.45.1
wrapt==1.17.2
zopfli==0.2.3.post1
